Refactor input classes to implement required interface.
