+++ src/Input/Contract/RuleHtmlByAttributeInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace PHPForge\Html\Input;
+namespace PHPForge\Html\Input\Contract;
 
 /**
  * Provide methods for handling HTML rule-related attributes from validators.
