+++ src/Input/Datetime.php
@@ -12,9 +12,10 @@ use PHPForge\Html\Attribute;
  *
  * @link https://www.w3.org/TR/2012/WD-html-markup-20120329/input.datetime.html#input.datetime
  */
-final class Datetime extends Base\AbstractInput
+final class Datetime extends Base\AbstractInput implements Contract\RequiredInterface
 {
     use Attribute\Custom\HasWidgetValidation;
+    use Attribute\Input\CanBeRequired;
     use Attribute\Input\HasMax;
     use Attribute\Input\HasMin;
     use Attribute\Input\HasStep;
